fix: move component in other folder
