+++ src/pages/HomePage.tsx
@@ -2,7 +2,7 @@ import SearchForm from "../components/components/SearchForm";
 import React from "react";
 import styled from "styled-components";
 import Header from "../components/Header";
-import CardList from "../components/CardList";
+import CardList from "../components/components/CardList";
 
 const StyledHomePage = styled.div`
   display: flex;
